pgc: wait for room in the error FIFO at the wrap
The full test compared a uint8_t write pointer against mapram[0x305] - 1,
which promotes to int and is -1 when the read pointer is 0, so it never
fired: the write pointer ran past 0xff, overwrote the buffer, and at every
wrap left wr == rd, which reads back as empty.

Both firmwares stall there instead. IBM's error writer (image 0x49e0) and
the IM-1024's (image 0x42e0) compare write+1 against the read pointer and,
when they match, spin on that same test until the host drains, servicing
only the display refresh. IBM Tech Ref p.81 agrees: "no writing may take
place until there is room". The loop's extra exits on C63FF, C6306 and
C6307 are kept, so 86Box stays more forgiving than the card.

fixes skiretic/im1024#30

+++ src/video/vid_pgc.c
@@ -214,7 +214,7 @@ static int
 error_byte(pgc_t *dev, uint8_t val)
 {
     /* If error buffer full, wait for it to empty. */
-    while (!dev->stopped && !dev->mapram[0x3ff] && !dev->mapram[0x306] && !dev->mapram[0x307] && dev->mapram[0x304] == dev->mapram[0x305] - 1) {
+    while (!dev->stopped && !dev->mapram[0x3ff] && !dev->mapram[0x306] && !dev->mapram[0x307] && dev->mapram[0x304] == (uint8_t) (dev->mapram[0x305] - 1)) {
         pgc_sleep(dev);
     }
